refactor: split slack-thread-processor.ts into focused modules
- Create 4 new converter modules in packages/dispatcher/src/converters/:
  - markdown-to-slack.ts: Markdown to Slack mrkdwn conversion
  - blockkit-processor.ts: Blockkit processing and action buttons
  - code-block-handler.ts: Code block parsing and validation
  - github-actions.ts: GitHub integration and PR buttons

- Refactor original slack-thread-processor.ts to use new modules
- Reduce main file from 1210 to 685 lines (43% reduction)
- Maintain backward compatibility and all existing functionality
- Follow single responsibility principle for better maintainability
